+++ cosmwasm/enclaves/shared/cosmos-proto/src/cosmwasm/msg.rs
@@ -478,7 +478,7 @@ pub struct MsgInstantiateContract {
     pub contract_id: ::std::string::String,
     pub init_msg: ::std::vec::Vec<u8>,
     pub auto_msg: ::std::vec::Vec<u8>,
-    pub init_funds: ::protobuf::RepeatedField<super::coin::Coin>,
+    pub funds: ::protobuf::RepeatedField<super::coin::Coin>,
     pub callback_sig: ::std::vec::Vec<u8>,
     pub contract_duration: ::std::string::String,
     // special fields
@@ -642,29 +642,29 @@ impl MsgInstantiateContract {
         ::std::mem::replace(&mut self.auto_msg, ::std::vec::Vec::new())
     }
 
-    // repeated .cosmos.base.v1beta1.Coin init_funds = 7;
+    // repeated .cosmos.base.v1beta1.Coin funds = 7;
 
 
-    pub fn get_init_funds(&self) -> &[super::coin::Coin] {
-        &self.init_funds
+    pub fn get_funds(&self) -> &[super::coin::Coin] {
+        &self.funds
     }
-    pub fn clear_init_funds(&mut self) {
-        self.init_funds.clear();
+    pub fn clear_funds(&mut self) {
+        self.funds.clear();
     }
 
     // Param is passed by value, moved
-    pub fn set_init_funds(&mut self, v: ::protobuf::RepeatedField<super::coin::Coin>) {
-        self.init_funds = v;
+    pub fn set_funds(&mut self, v: ::protobuf::RepeatedField<super::coin::Coin>) {
+        self.funds = v;
     }
 
     // Mutable pointer to the field.
-    pub fn mut_init_funds(&mut self) -> &mut ::protobuf::RepeatedField<super::coin::Coin> {
-        &mut self.init_funds
+    pub fn mut_funds(&mut self) -> &mut ::protobuf::RepeatedField<super::coin::Coin> {
+        &mut self.funds
     }
 
     // Take field
-    pub fn take_init_funds(&mut self) -> ::protobuf::RepeatedField<super::coin::Coin> {
-        ::std::mem::replace(&mut self.init_funds, ::protobuf::RepeatedField::new())
+    pub fn take_funds(&mut self) -> ::protobuf::RepeatedField<super::coin::Coin> {
+        ::std::mem::replace(&mut self.funds, ::protobuf::RepeatedField::new())
     }
 
     // bytes callback_sig = 8;
@@ -722,7 +722,7 @@ impl MsgInstantiateContract {
 
 impl ::protobuf::Message for MsgInstantiateContract {
     fn is_initialized(&self) -> bool {
-        for v in &self.init_funds {
+        for v in &self.funds {
             if !v.is_initialized() {
                 return false;
             }
@@ -757,7 +757,7 @@ impl ::protobuf::Message for MsgInstantiateContract {
                     ::protobuf::rt::read_singular_proto3_bytes_into(wire_type, is, &mut self.auto_msg)?;
                 },
                 7 => {
-                    ::protobuf::rt::read_repeated_message_into(wire_type, is, &mut self.init_funds)?;
+                    ::protobuf::rt::read_repeated_message_into(wire_type, is, &mut self.funds)?;
                 },
                 8 => {
                     ::protobuf::rt::read_singular_proto3_bytes_into(wire_type, is, &mut self.callback_sig)?;
@@ -795,7 +795,7 @@ impl ::protobuf::Message for MsgInstantiateContract {
         if !self.auto_msg.is_empty() {
             my_size += ::protobuf::rt::bytes_size(6, &self.auto_msg);
         }
-        for value in &self.init_funds {
+        for value in &self.funds {
             let len = value.compute_size();
             my_size += 1 + ::protobuf::rt::compute_raw_varint32_size(len) + len;
         };
@@ -829,7 +829,7 @@ impl ::protobuf::Message for MsgInstantiateContract {
         if !self.auto_msg.is_empty() {
             os.write_bytes(6, &self.auto_msg)?;
         }
-        for v in &self.init_funds {
+        for v in &self.funds {
             os.write_tag(7, ::protobuf::wire_format::WireTypeLengthDelimited)?;
             os.write_raw_varint32(v.get_cached_size())?;
             v.write_to_with_cached_sizes(os)?;
@@ -909,9 +909,9 @@ impl ::protobuf::Message for MsgInstantiateContract {
                 |m: &mut MsgInstantiateContract| { &mut m.auto_msg },
             ));
             fields.push(::protobuf::reflect::accessor::make_repeated_field_accessor::<_, ::protobuf::types::ProtobufTypeMessage<super::coin::Coin>>(
-                "init_funds",
-                |m: &MsgInstantiateContract| { &m.init_funds },
-                |m: &mut MsgInstantiateContract| { &mut m.init_funds },
+                "funds",
+                |m: &MsgInstantiateContract| { &m.funds },
+                |m: &mut MsgInstantiateContract| { &mut m.funds },
             ));
             fields.push(::protobuf::reflect::accessor::make_simple_field_accessor::<_, ::protobuf::types::ProtobufTypeBytes>(
                 "callback_sig",
@@ -945,7 +945,7 @@ impl ::protobuf::Clear for MsgInstantiateContract {
         self.contract_id.clear();
         self.init_msg.clear();
         self.auto_msg.clear();
-        self.init_funds.clear();
+        self.funds.clear();
         self.callback_sig.clear();
         self.contract_duration.clear();
         self.unknown_fields.clear();
@@ -971,7 +971,7 @@ pub struct MsgExecuteContract {
     pub contract: ::std::string::String,
     pub msg: ::std::vec::Vec<u8>,
     pub callback_code_hash: ::std::string::String,
-    pub sent_funds: ::protobuf::RepeatedField<super::coin::Coin>,
+    pub funds: ::protobuf::RepeatedField<super::coin::Coin>,
     pub callback_sig: ::std::vec::Vec<u8>,
     // special fields
     pub unknown_fields: ::protobuf::UnknownFields,
@@ -1093,29 +1093,29 @@ impl MsgExecuteContract {
         ::std::mem::replace(&mut self.callback_code_hash, ::std::string::String::new())
     }
 
-    // repeated .cosmos.base.v1beta1.Coin sent_funds = 5;
+    // repeated .cosmos.base.v1beta1.Coin funds = 5;
 
 
-    pub fn get_sent_funds(&self) -> &[super::coin::Coin] {
-        &self.sent_funds
+    pub fn get_funds(&self) -> &[super::coin::Coin] {
+        &self.funds
     }
-    pub fn clear_sent_funds(&mut self) {
-        self.sent_funds.clear();
+    pub fn clear_funds(&mut self) {
+        self.funds.clear();
     }
 
     // Param is passed by value, moved
-    pub fn set_sent_funds(&mut self, v: ::protobuf::RepeatedField<super::coin::Coin>) {
-        self.sent_funds = v;
+    pub fn set_funds(&mut self, v: ::protobuf::RepeatedField<super::coin::Coin>) {
+        self.funds = v;
     }
 
     // Mutable pointer to the field.
-    pub fn mut_sent_funds(&mut self) -> &mut ::protobuf::RepeatedField<super::coin::Coin> {
-        &mut self.sent_funds
+    pub fn mut_funds(&mut self) -> &mut ::protobuf::RepeatedField<super::coin::Coin> {
+        &mut self.funds
     }
 
     // Take field
-    pub fn take_sent_funds(&mut self) -> ::protobuf::RepeatedField<super::coin::Coin> {
-        ::std::mem::replace(&mut self.sent_funds, ::protobuf::RepeatedField::new())
+    pub fn take_funds(&mut self) -> ::protobuf::RepeatedField<super::coin::Coin> {
+        ::std::mem::replace(&mut self.funds, ::protobuf::RepeatedField::new())
     }
 
     // bytes callback_sig = 6;
@@ -1147,7 +1147,7 @@ impl MsgExecuteContract {
 
 impl ::protobuf::Message for MsgExecuteContract {
     fn is_initialized(&self) -> bool {
-        for v in &self.sent_funds {
+        for v in &self.funds {
             if !v.is_initialized() {
                 return false;
             }
@@ -1172,7 +1172,7 @@ impl ::protobuf::Message for MsgExecuteContract {
                     ::protobuf::rt::read_singular_proto3_string_into(wire_type, is, &mut self.callback_code_hash)?;
                 },
                 5 => {
-                    ::protobuf::rt::read_repeated_message_into(wire_type, is, &mut self.sent_funds)?;
+                    ::protobuf::rt::read_repeated_message_into(wire_type, is, &mut self.funds)?;
                 },
                 6 => {
                     ::protobuf::rt::read_singular_proto3_bytes_into(wire_type, is, &mut self.callback_sig)?;
@@ -1201,7 +1201,7 @@ impl ::protobuf::Message for MsgExecuteContract {
         if !self.callback_code_hash.is_empty() {
             my_size += ::protobuf::rt::string_size(4, &self.callback_code_hash);
         }
-        for value in &self.sent_funds {
+        for value in &self.funds {
             let len = value.compute_size();
             my_size += 1 + ::protobuf::rt::compute_raw_varint32_size(len) + len;
         };
@@ -1226,7 +1226,7 @@ impl ::protobuf::Message for MsgExecuteContract {
         if !self.callback_code_hash.is_empty() {
             os.write_string(4, &self.callback_code_hash)?;
         }
-        for v in &self.sent_funds {
+        for v in &self.funds {
             os.write_tag(5, ::protobuf::wire_format::WireTypeLengthDelimited)?;
             os.write_raw_varint32(v.get_cached_size())?;
             v.write_to_with_cached_sizes(os)?;
@@ -1293,9 +1293,9 @@ impl ::protobuf::Message for MsgExecuteContract {
                 |m: &mut MsgExecuteContract| { &mut m.callback_code_hash },
             ));
             fields.push(::protobuf::reflect::accessor::make_repeated_field_accessor::<_, ::protobuf::types::ProtobufTypeMessage<super::coin::Coin>>(
-                "sent_funds",
-                |m: &MsgExecuteContract| { &m.sent_funds },
-                |m: &mut MsgExecuteContract| { &mut m.sent_funds },
+                "funds",
+                |m: &MsgExecuteContract| { &m.funds },
+                |m: &mut MsgExecuteContract| { &mut m.funds },
             ));
             fields.push(::protobuf::reflect::accessor::make_simple_field_accessor::<_, ::protobuf::types::ProtobufTypeBytes>(
                 "callback_sig",
@@ -1322,7 +1322,7 @@ impl ::protobuf::Clear for MsgExecuteContract {
         self.contract.clear();
         self.msg.clear();
         self.callback_code_hash.clear();
-        self.sent_funds.clear();
+        self.funds.clear();
         self.callback_sig.clear();
         self.unknown_fields.clear();
     }
@@ -1355,7 +1355,7 @@ static file_descriptor_proto_data: &'static [u8] = b"\
     \tR\x10callbackCodeHash\x12#\n\x07code_id\x18\x03\x20\x01(\x04R\x06codeI\
     dB\n\xe2\xde\x1f\x06CodeID\x12\x1f\n\x0bcontract_id\x18\x04\x20\x01(\tR\
     \ncontractId\x12\x19\n\x08init_msg\x18\x05\x20\x01(\x0cR\x07initMsg\x12\
-    \x19\n\x08auto_msg\x18\x06\x20\x01(\x0cR\x07autoMsg\x12j\n\ninit_funds\
+    \x19\n\x08auto_msg\x18\x06\x20\x01(\x0cR\x07autoMsg\x12j\n\nfunds\
     \x18\x07\x20\x03(\x0b2\x19.cosmos.base.v1beta1.CoinR\tinitFundsB0\xaa\
     \xdf\x1f(github.com/cosmos/cosmos-sdk/types.Coins\xc8\xde\x1f\0\x122\n\
     \x0ccallback_sig\x18\x08\x20\x01(\x0cR\x0bcallbackSigB\x0f\xe2\xde\x1f\
@@ -1364,7 +1364,7 @@ static file_descriptor_proto_data: &'static [u8] = b"\
     \x06sender\x18\x01\x20\x01(\tR\x06sender\x12\x1a\n\x08contract\x18\x02\
     \x20\x01(\tR\x08contract\x12\x10\n\x03msg\x18\x03\x20\x01(\x0cR\x03msg\
     \x12,\n\x12callback_code_hash\x18\x04\x20\x01(\tR\x10callbackCodeHash\
-    \x12j\n\nsent_funds\x18\x05\x20\x03(\x0b2\x19.cosmos.base.v1beta1.CoinR\
+    \x12j\n\nfunds\x18\x05\x20\x03(\x0b2\x19.cosmos.base.v1beta1.CoinR\
     \tsentFundsB0\xaa\xdf\x1f(github.com/cosmos/cosmos-sdk/types.Coins\xc8\
     \xde\x1f\0\x122\n\x0ccallback_sig\x18\x06\x20\x01(\x0cR\x0bcallbackSigB\
     \x0f\xe2\xde\x1f\x0bCallbackSig:\x04\x88\xa0\x1f\0B3Z1github.com/trstlab\